Refactor NewsroomMain component and newsroomService
This commit refactors the NewsroomMain component and the newsroomService module. It updates the fetchArticles function in NewsroomMain to correctly set the currentArticles state and adds a formatDate function to format the date. In newsroomService, it replaces the deprecated getDoc function with getDocs and modifies the getAllNews function to return an array of articles instead of a single object.